requirements.txt
pandas>=1.5.0
openai>=1.0.0
datasets>=2.14.0
tqdm>=4.64.0
numpy>=1.24.0
Pillow>=9.0.0
openpyxl>=3.0.0
requests>=2.28.0
huggingface_hub>=0.16.0 